+++ NEWS.rst
@@ -6,12 +6,18 @@ every change, see the Git log.
 
 Latest
 ------
-* Minor: Added multicast examples.
+* Minor: Added multi-cast examples.
+* Major: Python object now only provide the functions they support. E.g., a non
+         trace encoder no longer has the trace function. Also the ``has_``
+         functions have been removed.
+         This includes:
+         ``has_partial_decoding_tracker``, ``has_systematic_encoder``,
+         and ``has_trace``.
 
 2.2.0
 -----
 * Minor: Added documentation for the python functions.
-* Minor: Added keyword arguements for the python functions.
+* Minor: Added keyword arguments for the python functions.
 * BugFix: Removed unused dependencies guage and tables.
 
 2.1.0